feat: Handle color already taken in join game
This commit adds the functionality to handle the case when a color is already taken in the join game process. It disables the corresponding color option and updates the image to indicate that the color is taken. It also displays an error message to the user, prompting them to choose another color.